added cancel train functionality
sets cancel status to true and active status to false

+++ railconnect-spring-boot/src/main/java/com/app/entities/TrainEntity.java
@@ -36,7 +36,7 @@ public class TrainEntity {
 
     @Id
     @Column(name = "train_number", unique = true)
-    private int trainNumber;
+    private long trainNumber;
 
     @Column(name = "train_name", length = 20)
     private String trainName;
